+++ myos/VirtPageAllocator
@@ -1,0 +1,28 @@
+#ifndef __VIRT_PAGE_ALLOCATOR__
+#define __VIRT_PAGE_ALLOCATOR__
+#include "PhysPageAllocator"
+#include "size.h"
+#define HHDM_BASE 0xFFFFFF0000000000ULL
+class VirtPageAllocator {
+private:
+    static inline void invlpg(void* addr);
+public:
+    PhysPageAllocator* phy_allocator;
+    static constexpr uint64_t P = 1ull << 0;
+    static constexpr uint64_t RW = 1ull << 1;
+    static constexpr uint64_t US = 1ull << 2;
+    static constexpr uint64_t PWT = 1ull << 3;
+    static constexpr uint64_t PCD = 1ull << 4;
+    static constexpr uint64_t A = 1ull << 5;
+    static constexpr uint64_t D = 1ull << 6;
+    static constexpr uint64_t PS = 1ull << 7;
+    static constexpr uint64_t G = 1ull << 8;
+
+    static constexpr uint64_t NX = 1ull << 63;
+    VirtPageAllocator();
+    void* pml4; // 커널 가상주소(= HHDM + PML4 물리)로 보관
+    void init(PhysPageAllocator* _phy_allocator);
+    uint64_t alloc_virt_page(uint64_t va, uint64_t pa, uint64_t flags);
+};
+
+#endif // __VIRT_PAGE_ALLOCATOR__
