+++ myos/msr.h
@@ -1,0 +1,15 @@
+#ifndef __MSTR_H__
+#define __MSTR_H__
+#include "size.h"
+static inline uint64_t rdmsr(uint32_t msr) {
+    uint32_t low, high;
+    __asm__ volatile ("rdmsr" : "=a"(low), "=d"(high) : "c"(msr));
+    return ((uint64_t)high << 32) | low;
+}
+
+static inline void wrmsr(uint32_t msr, uint64_t value) {
+    uint32_t low = (uint32_t)value;
+    uint32_t high = (uint32_t)(value >> 32);
+    __asm__ volatile ("wrmsr" : : "a"(low), "d"(high), "c"(msr));
+}
+#endif // __MSTR_H__
