+++ myos/size.h
@@ -1,0 +1,14 @@
+#ifndef __STDINT_H__
+#define __STDINT_H__
+
+typedef signed char        int8_t;
+typedef signed short       int16_t;
+typedef signed int         int32_t;
+typedef signed long long   int64_t;
+
+typedef unsigned char      uint8_t;
+typedef unsigned short     uint16_t;
+typedef unsigned int       uint32_t;
+typedef unsigned long long uint64_t;
+
+#endif /* __STDINT_H__ */
