+++ myos/trash.cpp_
